+++ R/prediction.R
@@ -1,3 +1,4 @@
+#' @rdname prediction
 #' @title Extract Predictions from a Model Object
 #' @description Extract predicted values via \code{\link[stats]{predict}} from a model object, conditional on data
 #' @param model A model object, perhaps returned by \code{\link[stats]{lm}} or \code{\link[stats]{glm}}.
@@ -13,7 +14,39 @@
 #' @seealso \code{\link{marginal_effects}}, \code{\link{build_margins}}
 #' @keywords models
 #' @export
-prediction <- function(model, data, type = c("response", "link"), ...) {
+prediction <- function(model, data, ...) {
+    UseMethod("prediction")
+}
+
+#' @rdname prediction
+#' @export
+prediction.lm <- function(model, data, type = "response", ...) {
+    # setup data
+    if (missing(data)) {
+        if (!is.null(model[["call"]][["data"]])) {
+            data <- eval(model[["call"]][["data"]], parent.frame()) 
+        } else { 
+            data <- get_all_vars(model[["terms"]], data = model[["model"]])
+        }
+    }
+    
+    type <- match.arg(type)
+    
+    # extract predicted value at input value (value can only be 1 number)
+    pred <- predict(model, newdata = data, type = type, se.fit = TRUE, ...)
+    class(pred[["fit"]]) <- c("fit", "numeric")
+    class(pred[["se.fit"]]) <- c("se.fit", "numeric")
+    
+    # obs-x-2 data.frame of predictions
+    structure(list(fitted = pred[["fit"]], 
+                   se.fitted = pred[["se.fit"]]), 
+              class = "data.frame", 
+              row.names = seq_len(length(pred[["fit"]])))
+}
+
+#' @rdname prediction
+#' @export
+prediction.glm <- function(model, data, type = c("response", "link"), ...) {
     # setup data
     if (missing(data)) {
         if (!is.null(model[["call"]][["data"]])) {